feat(catali): add boss mechanics and improve HUD
- Add boss unit configuration with respawn timer and experience multiplier
- Implement boss spawning logic based on team level with health scaling
- Update HUD to display boss health and respawn timer
- Change ExpGainEvent amount from int to float to support boss multipliers
- Move core healing to update loop and remove redundant method

+++ plugin/src/main/java/plugin/gamemode/catali/event/ExpGainEvent.java
@@ -8,7 +8,7 @@ import plugin.gamemode.catali.data.CataliTeamData;
 @RequiredArgsConstructor
 public class ExpGainEvent {
     public final CataliTeamData team;
-    public final int amount;
+    public final float amount;
     public final float x;
     public final float y;
 }
